Redo Pathfinding in order to take Terrain into account - change DrawRange to use a new Search method in gridMap. - add MovementScript that takes into account unit move range and type when getting range - change CalculateDistance in Pathfinding to use the MovementCostBook to determine the distance of a tile rather than just weighing each tile the same

+++ Assets/Scripts/View Model Component/TileInfo.cs
@@ -12,8 +12,7 @@ public class TileInfo
     [HideInInspector]
     public TileInfo prev;
     [HideInInspector]
-    public int distance;
-    
+    public float distance;
 
     public TileInfo(int x, int y, TerrainTypes terrain, GameObject content =null)
     {
